CurrentScope is incorrect
In the first iteration, a new scope will be created and pushed onto the top of the stack. CurrentScope is now a scope at the top of the stack, but in the second iteration, the newly added scope will change to the bottom. CurrentScope becomes the original Scope at the bottom of the stack.
My code only solved my problem, I didn't find the reason for stack inversion.