Trivial - fix URL for README badge
It was irking me that when I clicked on the badge to get
to the latest CI build, I'd get an error that the project didn't exist.
Fixes the URL to AppVeyor to see the latest build.

This also fixes the status image to the v2 branch.  Once v2 is in master
that querystring param fixing it to a branch should be ditched.