WIP #664 - De-duplicate in-use parts of YML
This is a bit of naughty scope-creep on that issue.
I noticed that there's some further parts of the YML which
are duplicated between the two v2 sections of the YML.

It seems that the intent is mainly about deploying to MyGet on
every build (in the appropriate branch) but only deploying to NuGet
if building from a tag.  Also, the actual build steps between these two
seem to be out of sync.

I've switched to a more concise way of achieving this same result,
which allows me to de-duplicate the two build configs so they're using
the same build process.

The only other real thing I did there was sneak in the autocrlf git
setting.  It's just a good idea on Windows to avoid being caught out.