WIP #665 - Beginnings of SonarCloud integration
Adds an XUnit test logger to each test csproj
file (used by test script).

The main things going on here in the YML are:
* Adding placeholders for some env vars for SonarCloud
* Adding the SonarScanner package as a build tool
* Adding before-build/after-test steps to setup/teardown SonarScanner
* Separating the "pack" step from the "build" step
  * Pack moved to before-package which occurs after tests pass
  * See https://www.appveyor.com/docs/build-configuration/#build-pipeline
* Tests run with logging & code-coverage detection
* Test results are added as AppVeyor artifacts

This YML won't work yet because at the very minimum the env
vars need to be filled in with real values.  This is the rough
gist of the SonarCloud config though.

Also add a skeleton SonarCloud analysis config file with a couple
of settings for starters: Don't analyse exception classes for coverage &
dont analyse samples for coverage.