Fix build failure: ForEachWorkflowTests
It seems that some work done in these commits, for the
file ForEachBuilderExtensions was responsible:
* 4807ace13c54f7395e322cd81e7395163e9f2b9c
* dfbba1e1762c870da1b6c258daed8a7cfc5e5423

A couple of the extension methods were missing out a ToList()
which later caused an ArgumentException when setting a property
value via reflection.

I also added an extra exception for the set-value logic.  This way,
if/when there is a problem, the exception will indicate which property
it was trying to set at the time.