Ensure that StartupRunner hosted service is registered before Elsa Options get evaluated
This prevents an issue when the user invokes StartWorkflow<T>, which would register the StartWorkflow<T> hosted service before the StartupRunner hosted service is registered.
The problem with that is that the migrations runner for EF Core executes *after* the StartWorkflow hosted service, which would then fail because the DB isn't prepared at that point.