Acquire lock on correlation ID when dispatching new workflows
This fixes a race condition between initial workflow execution that e.g. sends a message to a topic, which is then handled by the same workflow later on.
But if that message is delivered too soon (before the workflow got persisted & indexed), then the message consumer won't be able to find the workflow.

Acquiring a lock on the correlation ID fixes this by synchronizing access to that resource