Breaking behavior! Update OutcomeResult to NOT auto-connect to Done connections
This was originally added to support workflows created with the workflow builder API to automatically move on to the next activity. This is discouraged and one should instead always create an explicit connection.
The code being removed was causing a bug where the Done activity would always execute if there was a connection, even though the activity explicitly returned a different outcome.