Fix workflow selection on workflow instance viewer screen
This fixes the issue where the workflow dropdown list only displays workflow blueprints of the workflow instances currently displayed on screen.

It's not ideal, because if a user has hundreds or thousands of workflow definitions, this screen becomes basically useless.

A better implementation is to have a a combobox where the user can start typing in a workflow name (maybe initially showing the first 20 workflows)