Refactor workflow input handling
This update changes when workflow instances are created when querying workflows.

Before the update, new workflow instances would be created for startable workflows.
After this update, startable workflow instances only get saved if they are actually executed.

This change also breaks a few service signatures that originally accepted workflow input.
With this change, the caller is now expected to set the input explicitly on the workflow instance to execute.
This allows for better control when workflow instances need to be updated with workflow input.

Most notably, workflow input is now persisted with a workflow instance before it is executed or dispatched, allowing for increased odds of success when trying to retry a faulted workflow, since now it will have at least its input to draw upon.