Fix never-finishing workflow bug
This fixes an issue when a workflow tries to complete a workflow with the Finish activity, but never succeeds in doing so due to another scheduled activity creating a bookmark, causing the workflow to become suspended, rather than completed.

The fix is to have the Finish activity clear out the activity scheduler.