MongoDb Persistence: Cleanup and optout Serialzier
In this commit there is some cleanup of MongoDb persistence
provider initialization, as well as the option to opt out the automatic
registration of VariablesSerializer, to use default BsonSerializer
or custom version of BsonSerializer.

The option is opt-out, so previous code is unaffected.

Added also option to specify test mongodb connection string
from environment variable.

Related to #3538