Fix composite data access
The changed logic is necessary in order for custom activity type providers that yield composite activities to work. The problem is that the type name of a dynamically provided composite activity is a custom value, while the Type property always returns the class name. Comparing by activity ID fixes this.