Fall back to JsonObject
When deserialization dails due to the target type not being deserializable, return a JsonObject instead. This enable e.g. ActivityExecutionLogRecord to store JSON information for ActivityState, which does not need to be reconstructed into the original type. An improvement could be considered to store activity execution log record activity state in a non-polymorphic structure in the first place.